Create Appliances abstract
1.
Problem Statement:
You are tasked with designing a system to manage a set of appliances. Each appliance has common characteristics and behaviors, but their operation may differ based on the type of appliance. Use abstract classes to model this system.

Instructions:
Abstract Class Design

Create an abstract class named Appliance with the following:
Fields:
String brand
boolean isOn
Constructor to initialize brand and set isOn to false initially.
Abstract method operate(): Defines the specific operation of an appliance.
Concrete methods:
turnOn(): Sets isOn to true and prints "The appliance is now ON."
turnOff(): Sets isOn to false and prints "The appliance is now OFF."
Subclasses

Create two subclasses of Appliance:
WashingMachine:
Override the operate() method to print "The washing machine is washing clothes."
Refrigerator:
Override the operate() method to print "The refrigerator is cooling items."
Driver Class

Create a Main class to:
Instantiate a WashingMachine and a Refrigerator.
Turn on each appliance and call their operate() method.
Turn off each appliance.


O/p:
Washing Machine:
The appliance is now ON.
The washing machine is washing clothes.
The appliance is now OFF.

Refrigerator:
The appliance is now ON.
The refrigerator is cooling items.
The appliance is now OFF.